+++ packages/apollo-link-state/src/index.ts
@@ -6,6 +6,7 @@ import {
   FetchResult,
 } from 'apollo-link';
 import { ApolloCache } from 'apollo-cache';
+import { DocumentNode } from 'graphql';
 
 import { hasDirectives, getMainDefinition } from 'apollo-utilities';
 
@@ -14,7 +15,7 @@ const { graphql } = Async;
 
 import { FragmentMatcher } from 'graphql-anywhere';
 
-import { removeClientSetsFromDocument } from './utils';
+import { removeClientSetsFromDocument, normalizeTypeDefs } from './utils';
 
 const capitalizeFirstLetter = str => str.charAt(0).toUpperCase() + str.slice(1);
 
@@ -22,7 +23,7 @@ export type ClientStateConfig = {
   cache?: ApolloCache<any>;
   resolvers: any | (() => any);
   defaults?: any;
-  typeDefs?: string | string[];
+  typeDefs?: string | string[] | DocumentNode | DocumentNode[];
   fragmentMatcher?: FragmentMatcher;
 };
 
@@ -47,10 +48,8 @@ export const withClientState = (
     ): Observable<FetchResult> {
       if (typeDefs) {
         const directives = 'directive @client on FIELD';
-        const definition =
-          typeof typeDefs === 'string'
-            ? typeDefs
-            : typeDefs.map(typeDef => typeDef.trim()).join('\n');
+
+        const definition = normalizeTypeDefs(typeDefs);
 
         operation.setContext(({ schemas = [] }) => ({
           schemas: schemas.concat([{ definition, directives }]),
@@ -118,11 +117,18 @@ export const withClientState = (
               data: {},
             });
 
-      return obs.flatMap(
-        ({ data, errors }) =>
-          new Observable(observer => {
+      return new Observable(observer => {
+        // Works around race condition between completion and graphql execution
+        // finishing. If complete is called during the graphql call, we will
+        // miss out on the result, since the observer will have completed
+        let complete = false;
+        let handlingNext = false;
+        obs.subscribe({
+          next: ({ data, errors }) => {
             const observerErrorHandler = observer.error.bind(observer);
             const context = operation.getContext();
+
+            handlingNext = true;
             //data is from the server and provides the root value to this GraphQL resolution
             //when there is no resolver, the data is taken from the context
             graphql(resolver, query, data, context, operation.variables, {
@@ -133,11 +139,22 @@ export const withClientState = (
                   data: nextData,
                   errors,
                 });
-                observer.complete();
+                if (complete) {
+                  observer.complete();
+                }
+                handlingNext = false;
               })
               .catch(observerErrorHandler);
-          }),
-      );
+          },
+          error: observer.error.bind(observer),
+          complete: () => {
+            if (!handlingNext) {
+              observer.complete();
+            }
+            complete = true;
+          },
+        });
+      });
     }
   }();
 };