+++ packages/apollo-link-state/CHANGELOG.md
@@ -1,7 +1,8 @@
 # Change log
 
 ### vNEXT
-- Allow to provide resolvers via function
+- Allow providing resolvers via function [#293](https://github.com/apollographql/apollo-link-state/pull/293)
+- Add support for DocumentNode input to `typeDefs` [#284](https://github.com/apollographql/apollo-link-state/pull/284)
 
 ### 0.4.1
 - Return defaults when no Queries resolver is found and check for `resolverMap` [#202](https://github.com/apollographql/apollo-link-state/pull/202)